cmd/go: check missing go line in go.mod when 'go get go@1.N.P'
Currently, go will give a default go version to a go.mod which
doesn't have a go line. This causes confusing outputs when running
'go get go@1.22.1' as it treats the default go version as the current
version.

This CL ignores the default go line value of go.mod if the original
go.mod doesn't contain one, so it makes the output more precise.

Fixes: golang/go#70090

Change-Id: Ib9f16cd74adc4e93dfb547873f6cffe5f158dd25