cmd/compile,runtime: add column number(discriminator), plumb to pprof
Plumb the column number as addition to the line number to pprof. It is to be used as the discriminator for PGO basic block layout pass, but more usage of this field could be found.

These values are plumbed through runtime pc metadata and is ultimate
encoded in the pprof proto Line.Column field. It will be plumbed to
DWARF as well.

This makes the cmd/go binary 5% bigger,

For #59612.
For #65466.

Change-Id: Idb67de348de43b6d0f38d84c36e093a63d6b9c48