cmd/compile: implement basic block reordering based on edge profile
This CL added a frequency propagation pass after block frequency
assignment. It's a heuristic based data flow analysis, details are in
code comments. The pass is not guaranteed to converge.

This CL also added flags to enable this pass and an additional heuristic
in the layout pass to utilize the constructed edge profile. The hotter
edge will be scheduled first.

This CL is based on jinl@uber.com's cl/571535.

Change-Id: I65381f85809caab15b37538a56a56d5822a12562