internal/runtime/maps: medium map optimization
When a map is bigger than a group (8 entries) but smaller than a table
(1024 entries), we don't need a dictionary. Just point directly at the
single table containing all the data.

This saves some storage and means 1 less indrection to reach the data.

It does require a branch, but that branch exists already in directoryIndex.

Benchmarks don't see any difference. Saving the extra allocation is
probably worth it though. Not high priority if we don't get it in.

TODO: tell delve about this somehow?

Change-Id: I8637e2ddd9d64c5678473f4e7eac9bc85d1609d0