encoding/json: check exact structure of local error types
During the development of error wrapping (#29934),
this test was modified to stop using reflect.DeepEqual
since the prototype for error wrapping at the time included
frame information of where the error was created.

However, that change diminished the fidelity of the test
so that it is no longer as strict, which affects the endeavor
to implement v1 in terms of the v2 prototype.

For locally declared error types, use reflect.DeepEqual
to check that the exact structure is correct.

Change-Id: I443d418533866ab8d533bca3785fdc741e2c140e