crypto/tls: test with TLS FIPS mode
For tests that are interested in testing the difference between TLS in
FIPS mode or otherwise two new helpers are introduced,
runWithFIPSEnabled and runWithFIPSDisabled. They take care of forcing
the correct TLS FIPS state regardless of the overal GODEBUG=fips state,
and restoring it afterwards.

For the (many) tests that use features or test data not appropriate for
TLS in FIPS mode we add skips. For some tests we can make them
appropriate for both FIPS or not by tweaking some parameters that
weren't important to the subject under test, but would otherwise
preclude FIPS mode (e.g. because they used TLS 1.0 when the test could
use TLS 1.2 instead).

TODO:
* More tests could be run by regenerating some test certificates.
* More tests could (possibly) be run by updating golden test file data
  after adjusting ciphersuites, etc.

Change-Id: I311f3828dac890bb3ff8ebda6ed73d50f0797110