DO NOT SUBMIT: demonstrate rematerializable regalloc issue
This adds a useless no-op POR of the result of POPCNTQ with 0.

Build with `GOAMD64=v2 go build runtime`

runtime.sweepLocked.countAlloc contains a use of sys.OnesCount64. In
that function we can see the result of register allocation:

    (+1418) v47 = POPCNTQ <int> v46 : SI
    (-1418) v34 = Copy <int> v47 : X0
    (-1418) v38 = MOVQconst <int> [0] : X1
    (1418) v48 = POR <int> v34 v38 : X0

MOVQconst is assigned to X1 even though the regspec for MOVQconst does
not allow fp outputs.

Change-Id: If6340de95d6ce8ed978fff3ba9d9e40d13d0e2be