testing: annotate output text type
In the output of go test -json, this change provides a way to
disambiguate output produced by (*testing.T).Error{,f} and
(*testing.T).Fatal{,f} from other test logging. This allows test tooling
such as CI systems to identify which part of the output is most
pertinent for constructing summaries of test failures.

This is achieved by adding an OutputType field to output events. The
output type for an error is "error" for the first line and
"error-continue" for subsequent lines. The output type for framing is
"frame". This is achieved by bracketing error output with ^O and ^N,
escaped with ^[.

Fixes golang/go#62728.

Change-Id: Ib09c18ed5f729e1ae6d335cd1ec7d818c71532e0